students/seungjun/requirements.txt
Django==3.2.7

django-cors-headers==3.8.0

mysqlclient==2.0.3

bcrypt==3.2.0

PyJWT==2.1.0